+++ packages/evokit-box/src/index.js
@@ -1,3 +1,4 @@
+// 3.2.0-beta.1
 import { createBlock } from 'evokit';
 
 export const Box = createBlock('div', 'box', [